+++ src/main.rs
@@ -214,8 +214,8 @@ enum DepositCommands {
         destination_address: String,
         /// Fee rate to use for the recovery transaction (in sats/vB)
         fee_rate: u64,
-        /// Deposited output amount in BTC (e.g., 0.1 for 0.1 BTC)
-        amount: f64,
+        /// Deposited output amount in sats (e.g., 10000000 for 0.1 BTC)
+        amount: u64,
         /// Bitcoin network to use
         #[arg(long, default_value_t = CliNetwork::Bitcoin, value_enum)]
         network: CliNetwork,
@@ -228,9 +228,9 @@ enum DepositCommands {
         recovery_taproot_address: String,
         /// Citrea address (EVM address to receive bridged BTC)
         evm_address: String,
-        /// Deposited output amount in BTC (e.g., 0.1 for 0.1 BTC)
+        /// Deposited output amount in sats (e.g., 10000000 for 0.1 BTC)
         #[arg(long)]
-        amount: Option<f64>,
+        amount: Option<u64>,
         /// Bitcoin network to use
         #[arg(long, default_value_t = CliNetwork::Bitcoin, value_enum)]
         network: CliNetwork,
